Updates to snmptrap handling fix a bug in findDeviceByIP.  Add more tests for that. Move handle outside of the Trap class, it doesn't fit. Add developer docs.
