+++ LibreNMS/Snmptrap/Handlers/BgpBackwardTransition.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * LinkDown.php
+ * BgpBackwardTransition.php
  *
  * -Description-
  *
@@ -23,14 +23,14 @@
  * @author     Tony Murray <murraytony@gmail.com>
  */
 
-namespace LibreNMS\Snmptrap\Handler;
+namespace LibreNMS\Snmptrap\Handlers;
 
 use App\Models\Device;
 use LibreNMS\Interfaces\SnmptrapHandler;
 use LibreNMS\Snmptrap\Trap;
 use Log;
 
-class LinkDown implements SnmptrapHandler
+class BgpBackwardTransition implements SnmptrapHandler
 {
     /**
      * Handle snmptrap.
@@ -42,30 +42,22 @@ class LinkDown implements SnmptrapHandler
      */
     public function handle(Device $device, Trap $trap)
     {
-        $ifIndex = $trap->getOidData($trap->findOid('IF-MIB::ifIndex'));
+        $state_oid = $trap->findOid('BGP4-MIB::bgpPeerState');
+        $bgpPeerIp = substr($state_oid, 23);
 
-        $port = $device->ports()->where('ifIndex', $ifIndex)->first();
+        $bgpPeer = $device->bgppeers()->where('bgpPeerIdentifier', $bgpPeerIp)->first();
 
-        if (!$port) {
-            Log::warning("Snmptrap linkDown: Could not find port at ifIndex $ifIndex for device: " . $device->hostname);
+        if (!$bgpPeer) {
+            Log::error('Unknown bgp peer handling bgpEstablished trap: ' . $bgpPeerIp);
             return;
         }
 
-        $port->ifOperStatus = $trap->getOidData("IF-MIB::ifOperStatus.$ifIndex");
-        $port->ifAdminStatus = $trap->getOidData("IF-MIB::ifAdminStatus.$ifIndex");
+        $bgpPeer->bgpPeerState = $trap->getOidData($state_oid);
 
-        $device_array = $device->toArray();
-
-        log_event("SNMP Trap: linkDown $port->ifAdminStatus/$port->ifOperStatus " . $port->ifDescr, $device_array, 'interface', 5, $port->port_id);
-
-        if ($port->isDirty('ifAdminStatus')) {
-            log_event("Interface Disabled : " . $port['ifDescr'] . " (TRAP)", $device_array, "interface", 3, $port['port_id']);
+        if ($bgpPeer->isDirty('bgpPeerState')) {
+            log_event('SNMP Trap: BGP Down ' . $bgpPeer->bgpPeerIdentifier . ' ' . get_astext($bgpPeer->bgpPeerRemoteAs) . ' is now ' . $bgpPeer->bgpPeerState, $device->toArray(), 'bgpPeer', 5, $bgpPeerIp);
         }
 
-        if ($port->isDirty('ifOperStatus')) {
-            log_event("Interface went Down : " . $port['ifDescr'] . " (TRAP)", $device_array, "interface", 5, $port['port_id']);
-        }
-
-        $port->save();
+        $bgpPeer->save();
     }
 }
